+++ app/src/main/java/hu/tb/minichefy/presentation/screens/recipe_list/navigation/RecipeListNavigation.kt
@@ -13,7 +13,7 @@ fun NavController.navigateToRecipeList() {
 
 fun NavGraphBuilder.recipeListNavigation(
     navigateToCreateRecipe: () -> Unit,
-    navigateToRecipeDetails: (Int) -> Unit,
+    navigateToRecipeDetails: (Long) -> Unit,
 ) {
     composable(route = RECIPE_LIST_ROUTE) {
         RecipeListScreen(
